requirements.txt
Django==3.2.5
djangorestframework==3.12.4
Pillow==8.3.1
django-allauth==0.45.0
dj-rest-auth==2.1.8
django-cors-headers==3.7.0
